+++ test/specs/transducers.js
@@ -1,7 +1,7 @@
 import { expect } from "../helper";
 
 import { chan, CLOSED } from "modules/channel";
-import { go } from "modules/ops";
+import { go } from "modules/process";
 
 import { transducers as t } from "xduce";